requirements.txt
django >3.2,<3.3
python-decouple
pillow
djangorestframework
drf-yasg[validation]
elasticsearch-dsl==7.4.0
elasticsearch==7.10.1
django-elasticsearch-dsl==7.2.0
django-elasticsearch-dsl-drf==0.22.0
setuptools
celery
redis
django-redis
django-mail-templated
django-cors-headers
djangorestframework-simplejwt
requests
black
flake8